Initial version of single-document overview of ChatExchange API.
This replaces the getting started information in the README.

This describes every element of the intended public API. Things
like the browser module, _util module, and various methods that
are intended only for internal use (even if they're not
explicitly marked as such) are not included. It also includes
several examples and installation instructions.